# customerservice/main.go
package main

// Swagger docs
import (
	common "customerservice/common"
	database "customerservice/database"
	docs "customerservice/docs" // generated by swag
	userservice "customerservice/user"

	"github.com/gin-gonic/gin"
	log "github.com/sirupsen/logrus"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

// @title Customer Details API
// @version 1.0
// @description API for Customer service
// @host localhost:3003
// @BasePath /api
// @securityDefinitions.apikey Bearer
// @in header
// @name Authorization
// @description Type "Bearer" followed by a space and JWT token.
func main() {
	err := common.ConfigSetup("configuration/dbconfig.yaml")
	if err != nil {
		log.Error("ConfigSetup failed")
		return
	}

	configuration := common.GetConfig()
	err = database.SetupDB(configuration)

	if err != nil {
		log.Error("SetupDB failed")
		return
	} else {
		log.Info("DB Setup Success")
	}

	router := gin.Default()

	// Add CORS middleware
	router.Use(func(c *gin.Context) {
		c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, PATCH")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	})

	docs.SwaggerInfo.Description = "API for customer service"
	docs.SwaggerInfo.Version = "1.0"
	docs.SwaggerInfo.Host = ":3000"
	docs.SwaggerInfo.BasePath = "/api"
	docs.SwaggerInfo.Schemes = []string{"http"}

	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	// Public routes
	router.POST("/api/customersignup", userservice.AddNewCustomer)
	router.POST("/api/customerlogin", userservice.CustomerLogin)

	router.Run(":3000")
}